- complete rework

+++ LTB-Database/Core/DataModel/Book.cs
@@ -7,10 +7,10 @@ namespace LTB_Database.Core.DataModel
 {
 	public class Book
 	{
-		public int id { get; set; }
+		public long id { get; set; }
 		public string name { get; set; }
-		public int number { get; set; }
-		public int catid { get; set; }
+		public long number { get; set; }
+		public long catid { get; set; }
 		public string image { get; set; }
 		public DateTime added { get; set; }
 		public string category { get; set; }
